cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 kadeng muchulee8 aakhundov ColinPeppler amjames desertfire chauhang

[ghstack-poisoned]

+++ aten/src/ATen/ParallelCommon.cpp
@@ -107,7 +107,7 @@ int intraop_default_num_threads() {
     nthreads = TaskThreadPoolBase::defaultNumThreads();
 #endif
   }
-  return static_cast<int>(nthreads);
+  return nthreads;
 #endif
 }
