+++ c10/util/SmallBuffer.h
@@ -55,10 +55,11 @@ class SmallBuffer {
       delete[] data_;
     }
   }
-  T& operator[](size_t idx) {
+
+  T& operator[](int64_t idx) {
     return data()[idx];
   }
-  const T& operator[](size_t idx) const {
+  const T& operator[](int64_t idx) const {
     return data()[idx];
   }
   T* data() {
